requirements.txt
Adafruit-Blinka==3.9.0
adafruit-circuitpython-neopixel==5.0.0
Adafruit-PlatformDetect==2.2.0
Adafruit-PureIO==1.0.4
audioread==2.1.8
cffi==1.14.0
colorzero==1.1
decorator==4.4.1
gpiozero==1.5.1
joblib==0.14.1
librosa==0.7.2
llvmlite==0.31.0
numba==0.48.0
numpy==1.18.1
Pillow==7.0.0
py-midi==2.0.1
pycparser==2.19
pyftdi==0.43.0
pyserial==3.4
pyusb==1.0.2
resampy==0.2.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
SoundFile==0.10.3.post1
spidev==3.4
sysv-ipc==1.0.1
